🎨 start cleaning displayPage
- started cleaning displayPage so it becomes more modular
- started swapping out pure css to tailwind
