+++ srcs/server/frontend/ts/formHandlers.ts
@@ -2,7 +2,7 @@ import * as utils from './utils.js';
 import * as displayPage from './displayPage.js';
 import * as buttonHandlers from './buttonHandlers.js';
 import { globalSocket, initGlobalChat, onlineUsers } from './chatManager.js';
-import { getUnreadMessages } from './utils.js';
+import { getUnreadMessages } from './chatManager.js';
 
 export async function signUp(e: Event) {
 	e.preventDefault();
