+++ srcs/server/frontend/ts/main.ts
@@ -1,7 +1,8 @@
 import * as displayPage from './displayPage.js';
 import * as buttonHandlers from './buttonHandlers.js';
 import { initGlobalChat } from './chatManager.js';
-import { getUnreadMessages, initAppNav } from './utils.js';
+import { initAppNav } from './utils.js';
+import { getUnreadMessages } from './chatManager.js';
 
 async function isSignedIn() {
 	try {
